unify unit and integration tests on core-modules
