+++ packages/core-modules/.gitignore
@@ -1,4 +1,2 @@
-artifacts
-cache
 contracts/Router.sol
 deployments
