+++ packages/core-modules/.solcover.js
@@ -1,0 +1,3 @@
+module.exports = {
+  skipFiles: ['contracts/Proxy.sol', 'contracts/Router.sol'],
+};
