+++ packages/core-modules/hardhat.config.js
@@ -1,9 +1,12 @@
 require('solidity-coverage');
 require('@nomiclabs/hardhat-ethers');
+require('@synthetixio/deployer');
 
 module.exports = {
   solidity: '0.8.4',
-  paths: {
-    tests: './test/unit',
+  networks: {
+    local: {
+      url: 'http://localhost:8545',
+    },
   },
 };
