+++ packages/core-modules/test/contracts/modules/CoreUpgradeModule.test.js
