--- packages/core-modules/test/integration/.solcover.js
@@ -1,3 +1,0 @@
-module.exports = {
-  skipFiles: ['mocks', 'Router.sol'],
-};
