--- packages/core-modules/test/integration/contracts/modules/SampleModule.sol
@@ -1,9 +1,0 @@
-//SPDX-License-Identifier: MIT
-pragma solidity ^0.8.0;
-
-import "@synthetixio/core-modules/contracts/modules/CoreSampleModule.sol";
-
-// solhint-disable-next-line no-empty-blocks
-contract SampleModule is CoreSampleModule {
-
-}
