--- packages/core-modules/test/integration/hardhat.config.js
@@ -1,7 +1,0 @@
-require('solidity-coverage');
-require('@nomiclabs/hardhat-ethers');
-require('@synthetixio/deployer');
-
-module.exports = {
-  solidity: '0.8.4',
-};
